requirements.txt
Django==1.10
django-filter==1.0.4
djangorestframework==3.6.4
projectycopg2==2.7.3

